add module services

+++ src/main/java/com/nguyenvanancodeweb/lakesidehotel/repository/ServicesRepository.java
@@ -3,6 +3,6 @@ package com.nguyenvanancodeweb.lakesidehotel.repository;
 import com.nguyenvanancodeweb.lakesidehotel.model.Services;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServicesRepository extends JpaRepository<Services, Integer> {
+public interface ServicesRepository extends JpaRepository<Services, Long> {
 
 }
